d3-scripts/pyproject.toml
[tool.poetry]
name = "d3_scripts"
version = "0.1.0"
description = "Utility scripts for ManySecured-D3 claims"
authors = [
    "NquiringMinds <contact@nquiringminds.com>",
    "Alois Klink <alois@nquiringminds.com>",
    "Craig Rafter <craig@nquiringminds.com>"
]
readme = "README.md"
homepage = "https://github.com/TechWorksHub/ManySecured-D3DB"
repository = "https://github.com/TechWorksHub/ManySecured-D3DB.git"
packages = [
    { include = "d3_scripts", from = "src" }
]

[tool.poetry.dependencies]
python = ">=3.7,<4"
PyYAML = "^6.0"
requests = "^2.27.1"
jsonschema = "^4.4.0"
yamllint = "^1.26.3"
tqdm = "^4.64.0"
validators = "^0.19.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"

[tool.poetry.scripts]
d3lint = "d3_scripts.d3_lint:cli"
d3build = "d3_scripts.d3_build:d3_build"
d3export = "d3_scripts.d3_build_db:d3_build_db"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
